feat: include fee bump in proof send modal
